requirements.txt
fastAPI
requests
uvicorn
pydantic
pandas
python-jose
passlib
pytest
pytest-cov
sqlalchemy
psycopg2-binary
pylint